third_party/blink/web_tests/fast/forms/select/customizable-select/disallowed-select-descendants-console-message-expected.txt

CONSOLE ERROR: A descendant of a <select> does not follow the content model.
CONSOLE ERROR: A descendant of a <select> does not follow the content model.
CONSOLE ERROR: A descendant of a <select> does not follow the content model.
CONSOLE ERROR: A descendant of a <select> does not follow the content model.
CONSOLE ERROR: A descendant of a <select> does not follow the content model.
  
 
